+++ packages/plugin-code-block/web/src/editor.ts
@@ -2,12 +2,16 @@ import { createCodeBlockPlugin } from './createCodeBlockPlugin';
 import { CODE_BLOCK_TYPE, CodeBlockPluginEditorConfig } from './types';
 import { DEFAULTS } from './defaults';
 import { EditorPluginCreator } from 'wix-rich-content-common';
+import { createTiptapExtensions } from './tiptap';
+import { TiptapEditorPlugin } from 'ricos-tiptap-types';
 
 export const pluginCodeBlock: EditorPluginCreator<CodeBlockPluginEditorConfig> = config => {
+  const pluginConfig = { ...DEFAULTS.config, ...config };
   return {
-    config: { ...DEFAULTS.config, ...config },
+    config: pluginConfig,
     type: CODE_BLOCK_TYPE,
     createPlugin: createCodeBlockPlugin,
     ModalsMap: {},
-  };
+    tiptapExtensions: createTiptapExtensions(pluginConfig),
+  } as TiptapEditorPlugin;
 };
